Folder is cleaned up. UpdateIO now incorporates all MCP23017 functionalities. Now it is no longer needed to include mcp23017.h. These files also no longer exits anymore. Including io.h gives acces to mcpWrite() and mcpRead().
The yEd state diagrams which are to be used for state machines are to be stored in the folders: mainStateMachines and nestedStateMachines. All state machines will be generated when the
'assembleProject.py' script is executed. Afterwards the mainStateMachine and nestedStateMachineFolders will be empty and ready for reuse. The state diagram files are copied to yEd_stateMachinesRepo and
moved to the new project. So the new project will have the original yEd files in the stateDiagram sub directory.

The project is 100% compilable with the exception of undefined timers of nested state machines.

the src sub directory also has been added to support the Arduino software. This is needed to compile sub directories

The the generated project directory is not tested yet. At this point it is not known if all mcp and io functions actually work, they should though.