main state machines need their beginState defined in order to compiles. Nested state machines default to idle as they need to be initialized before usage. I2C initialisation was not done, it is now.
So far I have not been able to include more than 1 tab and/or spaces for the .tab files so users could outline these files.

The module picker is also added. Users are prompted to include modules. The mcp23017 module is deleted as it is absorbed in the IO files.

The basic is now finished. All code is so far is properly organized.

Now modules have to be made ready. The first must be the buttonClass (rename to debounceClass).
new modules must be made for:
- software pwm servo
- pca9685 pwm driver
- keypad
- whatever more comes to mind