added a more universal version of the short circuit function. This function may automatically re-apply trackpower 5 seconds after a short circuit. To let this happen the last argument of the init function must be true;
The module calculates the maxium ADC value based upon the maximum given milli Amps and milli Ohms of
the shunt resistor.

The function short circuit also needs a 1ms timer. To let this work, one must pass the address of one
in the init function.

The init function also need the track power pin number as well as the analog input pin number

I have not tried to compile this module let alone test it.